refactor(ffe-searchable-dropdown-react): rename folders to lower case
