+++ packages/ffe-searchable-dropdown-react/src/multi/MultiselectOptionBody.tsx
