+++ packages/ffe-searchable-dropdown-react/src/multi/getNewList.ts
