+++ packages/ffe-searchable-dropdown-react/src/single/OptionBody.tsx
