+++ packages/ffe-searchable-dropdown-react/src/single/reducer.ts
